pyproject.toml
[project]
name = "instagram-agent"
version = "0.1.0"
description = "AI agent for Instagram content strategy and photo management"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "instructor>=1.0.0",
    "crewai>=0.80.0",
    "tavily-python>=0.5.0",
    "celery[redis]>=5.3.0",
    "python-multipart>=0.0.12",
    "pillow>=11.0.0",
    "rawpy>=0.23.0",
    "apify-client>=1.7.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
